Access Structure field within an Array.
